refactor: run tests
